requirements.txt
Flask==1.0.3
Flask-Cors==3.0.7
pyOpenSSL==19.0.0
Flask-SQLAlchemy==2.4.0
SQLAlchemy==1.3.3
PyMySQL==0.9.3
pymysql-pooling==1.0.1
PyJWT==1.7.1
